+++ src/heap.jl
@@ -1,25 +1,25 @@
-@inline function heapify!(f, xs::Vector)
+@inline function heapify!(o::Ordering, xs::Vector)
     for i in div(length(xs), 2):-1:1
-        percolate_down!(f, xs, i, xs[i])
+        percolate_down!(o, xs, i, xs[i])
     end
     return xs
 end
 
 @inline function percolate_down!(
-        f::F,
+        o::Ordering,
         xs::Vector,
         i::Integer,
         x,
-    ) where {F}
+    )
     len = length(xs)
     @inbounds while (l = 2i) <= len
         r = 2i + 1
-        j = if r > len || f(last(@inbounds(xs[l])), last(@inbounds(xs[r])))
+        j = if r > len || lt(o, last(@inbounds(xs[l])), last(@inbounds(xs[r])))
             l
         else
             r
         end
-        if f(last(@inbounds(xs[j])), last(x))
+        if lt(o, last(@inbounds(xs[j])), last(x))
             @inbounds xs[i] = xs[j]
             i = j
         else
@@ -29,20 +29,20 @@ end
     return @inbounds xs[i] = x
 end
 
-@noinline function heappop!(f, xs::Vector)
+@noinline function heappop!(o::Ordering, xs::Vector)
     isempty(xs) && throw(BoundsError(xs, 1))
     x = @inbounds xs[1]
     y = @inbounds pop!(xs)
     if !isempty(xs)
-        percolate_down!(f, xs, 1, y)
+        percolate_down!(o, xs, 1, y)
     end
     return x
 end
 
-@inline function heapreplace!(f, xs::Vector, x)
+@inline function heapreplace!(o::Ordering, xs::Vector, x)
     @boundscheck isempty(xs) && throw(BoundsError(xs, 1))
     res = @inbounds xs[1]
     @inbounds xs[1] = x
-    percolate_down!(f, xs, 1, x)
+    percolate_down!(o, xs, 1, x)
     return res
 end